+++ network/base/templatetags/tags.py
@@ -18,5 +18,13 @@ def frq(value):
     except (TypeError, ValueError):
         return '-'
     formatted = format(float(to_format) / 1000000, '.3f')
-    formatted = formatted + ' Mhz'
+    formatted = formatted + ' MHz'
     return formatted
+
+
+@register.filter
+def percentagerest(value):
+    try:
+        return 100 - value
+    except (TypeError, ValueError):
+        return 0