+++ network/base/urls.py
@@ -1,6 +1,7 @@
 from django.conf.urls import patterns, url
 from django.views.generic import TemplateView
 
+from network.base.views import StationAllView
 
 urlpatterns = patterns(
     'network.base.views',
@@ -12,6 +13,7 @@ urlpatterns = patterns(
     # Observations
     url(r'^observations/$', 'observations_list', name='observations_list'),
     url(r'^observations/(?P<id>[0-9]+)/$', 'observation_view', name='observation_view'),
+    url(r'^observations/(?P<id>[0-9]+)/delete/$', 'observation_delete', name='observation_delete'),
     url(r'^observations/new/$', 'observation_new', name='observation_new'),
     url(r'^prediction_windows/(?P<sat_id>[\w.@+-]+)/(?P<start_date>.+)/(?P<end_date>.+)/$',
         'prediction_windows', name='prediction_windows'),
@@ -19,5 +21,8 @@ urlpatterns = patterns(
     # Stations
     url(r'^stations/$', 'stations_list', name='stations_list'),
     url(r'^stations/(?P<id>[0-9]+)/$', 'station_view', name='station_view'),
+    url(r'^stations/(?P<id>[0-9]+)/delete/$', 'station_delete', name='station_delete'),
     url(r'^stations/edit/$', 'station_edit', name='station_edit'),
+    url(r'^stations_all/$', StationAllView.as_view({'get': 'list'}), name='stations_all'),
+
 )