+++ Changelog.md
@@ -1,3 +1,15 @@
+## 0.5
+
+* Hide satellites with no transmitters
+* Success rate for Ground Stations
+* Sync MODES from DB
+* Allow user to delete Scheduled Observations
+* Add ability to delete a Ground Station
+* Add initialize command to help developers
+* Add Discussion button on Observations
+* Exclude overlapped timeframe from observations
+* Compress static files
+
 ## 0.4
 
 * Rename Transponders to Transmitters